refactor: avoid duplicate code in oai_sample
move ingest code to ingest module
